+++ packages/clerk-js/src/utils/web3.ts
@@ -1,23 +1,23 @@
 import type { Web3Provider } from '@clerk/types';
 
 import { toHex } from './hex';
+import { injectedWeb3Providers } from './injectedWeb3Providers';
 
 type GetWeb3IdentifierParams = {
   provider: Web3Provider;
 };
 
-export async function getWeb3Identifier(_: GetWeb3IdentifierParams): Promise<string> {
-  // @ts-ignore
-  if (!global.ethereum) {
-    // Do nothing when ethereum doesn't exist.
+export async function getWeb3Identifier(params: GetWeb3IdentifierParams): Promise<string> {
+  const { provider } = params;
+  const ethereum = await getEthereumProvider(provider);
+  if (!ethereum) {
+    // If a plugin for the requested provider is not found,
+    // the flow will fail as it has been the expected behavior so far.
     return '';
   }
 
+  const identifiers = await ethereum.request({ method: 'eth_requestAccounts' });
   // @ts-ignore
-  const identifiers = await global.ethereum.request({
-    method: 'eth_requestAccounts',
-  });
-
   return (identifiers && identifiers[0]) || '';
 }
 
@@ -27,15 +27,16 @@ type GenerateWeb3SignatureParams = {
   provider: Web3Provider;
 };
 
-export async function generateWeb3Signature({ identifier, nonce }: GenerateWeb3SignatureParams): Promise<string> {
-  // @ts-ignore
-  if (!global.ethereum) {
-    // Do nothing when ethereum doesn't exist.
+export async function generateWeb3Signature(params: GenerateWeb3SignatureParams): Promise<string> {
+  const { identifier, nonce, provider } = params;
+  const ethereum = await getEthereumProvider(provider);
+  if (!ethereum) {
+    // If a plugin for the requested provider is not found,
+    // the flow will fail as it has been the expected behavior so far.
     return '';
   }
 
-  // @ts-ignore
-  return await global.ethereum.request({
+  return await ethereum.request({
     method: 'personal_sign',
     params: [`0x${toHex(nonce)}`, identifier],
   });
@@ -49,6 +50,10 @@ export async function getCoinbaseIdentifier(): Promise<string> {
   return await getWeb3Identifier({ provider: 'coinbase' });
 }
 
+export async function getCoinbaseWalletIdentifier(): Promise<string> {
+  return await getWeb3Identifier({ provider: 'coinbase_wallet' });
+}
+
 type GenerateSignatureParams = {
   identifier: string;
   nonce: string;
@@ -60,4 +65,20 @@ export async function generateSignatureWithMetamask({ identifier, nonce }: Gener
 
 export async function generateSignatureWithCoinbase({ identifier, nonce }: GenerateSignatureParams): Promise<string> {
   return await generateWeb3Signature({ identifier, nonce, provider: 'coinbase' });
+}
+
+export async function generateSignatureWithCoinbaseWallet({
+  identifier,
+  nonce,
+}: GenerateSignatureParams): Promise<string> {
+  return await generateWeb3Signature({ identifier, nonce, provider: 'coinbase_wallet' });
+}
+
+async function getEthereumProvider(provider: Web3Provider) {
+  if (provider === 'coinbase_wallet') {
+    const CoinbaseWalletSDK = await import('@coinbase/wallet-sdk').then(mod => mod.CoinbaseWalletSDK);
+    const sdk = new CoinbaseWalletSDK({});
+    return sdk.makeWeb3Provider({ options: 'all' });
+  }
+  return injectedWeb3Providers.get(provider);
 }