+++ packages/testing/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @clerk/testing
 
+## 1.2.17
+
+### Patch Changes
+
+- Updated dependencies [[`82593173a`](https://github.com/clerk/javascript/commit/82593173aafbf6646e12c5779627cdcb138a1f27), [`afad9af89`](https://github.com/clerk/javascript/commit/afad9af893984a19d7284f0ad3b36e7891d0d733)]:
+  - @clerk/types@4.18.0
+  - @clerk/backend@1.9.1
+  - @clerk/shared@2.6.1
+
 ## 1.2.16
 
 ### Patch Changes